Add files via upload
We have a very large data without labels. So, we have to manually create the labels and it is not completed yet. So, we have used a very small part of the data to test the code but we have to create the labels manually and then finetune the model completely and then we will proceed with the deep learning models